% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinamica_utils.r
\name{dinamica_utils}
\alias{dinamica_utils}
\alias{check_dinamica}
\alias{get_dinamica_path}
\title{Dinamica Utility Functions}
\usage{
check_dinamica()

get_dinamica_path()
}
\description{
Various functions to interact with Dinamica from R
}
\section{Functions}{
\itemize{
\item \code{check_dinamica()}: Check Dinamica is installed

\item \code{get_dinamica_path()}: Get Dinamica path

}}
