% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_zenodo_predictors.r
\name{fetch_zenodo_predictors}
\alias{fetch_zenodo_predictors}
\title{Get predictor data from Zenodo}
\usage{
fetch_zenodo_predictors(
  url = "https://zenodo.org/records/8263509/files/LULCC_CH_dat.zip",
  target_dir = "data/preds/"
)
}
\arguments{
\item{url}{The url of the zip to fetch}

\item{target_dir}{The dir in which to depose the data}
}
\description{
This fetches predictor data from a very specifically structured Zenodo repo, namely
like \url{https://doi.org/10.5281/zenodo.8263509}
}
