% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{get_config}
\alias{get_config}
\title{LULCC_CH setup: control table, config}
\usage{
get_config(
  scenario_names = c("BAU", "EI-NAT", "EI-CUL", "EI-SOC", "GR-EX"),
  step_length = 5L,
  simctrl_tbl_path = "tools/simulation_control.csv"
)
}
\description{
Creates / reads control table and config files
}
