% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulc_data_prep.r
\name{lulc_data_prep}
\alias{lulc_data_prep}
\title{LULC_data_prep}
\usage{
lulc_data_prep()
}
\description{
Preparing land use land cover rasters for all historic periods in the the Swiss Areal
Statistiks and aggregating LULC rasters to new classification scheme
}
\author{
Ben Black
}
