% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.analysedownsampling.r
\name{lulcc.analysedownsampling}
\alias{lulcc.analysedownsampling}
\title{lulcc.analysedownsampling}
\usage{
lulcc.analysedownsampling(Comparative_table, summary_metrics)
}
\description{
Produce plots and perform statistical analysis to compare downsampling vs. regular RF
}
\author{
Ben Black
}
