% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.boyceroccurves.r
\name{lulcc.BoyceROCcurves}
\alias{lulcc.BoyceROCcurves}
\title{lulcc.BoyceAUCcurves}
\usage{
lulcc.BoyceROCcurves(f, pa, tesdat, crit, tre = numeric())
}
\arguments{
\item{f}{Vector of model predicted probabilities}

\item{pa}{Vector of response variable values}

\item{tesdat}{Dataframe of test data}

\item{crit}{which threshold criterion should be considered? Currently 'pp=op'
(predicted prevalence = observed prevalence), 'maxTSS' (threshold yielding maximum TSS),
and 'external' (thresholds manually supplied) are possible}

\item{tre}{Optional numeric threshold for seperating predictions}
}
\value{
List of model eval result objects: AUC, Boyce and the ROC value
}
\description{
Return objects for producing AUC and Boyce Curves
}
\details{
Not to be called directly by the user
}
\author{
Philipp Brun (main) with edits by Ben Black
}
