% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.evaluate.r
\name{lulcc.evaluate}
\alias{lulcc.evaluate}
\title{lulcc.evaluate models}
\usage{
lulcc.evaluate(
  x,
  tester = data.frame(),
  thres = numeric(),
  crit = "pp=op",
  prevalence_correction = FALSE,
  train_test
)
}
\arguments{
\item{x}{a wsl.fit object}

\item{tester}{data.frame with testing data (only mandatory if replicatetype='none'
was chosen when models were fitted)}

\item{crit}{which threshold criterion should be considered? Currently 'pp=op'
(predicted prevalence = observed prevalence), 'maxTSS' (threshold yielding maximum TSS),
and 'external' (thresholds manually supplied) are possible}

\item{prevalence_correction}{logical. Should imbalanced presence/absence data be upsampled to
prevalence 0.5 for model evaluation.
data. In a pseudoabsence correction the same amount of pseudoabsences as presences are
subsampled for model validation.}

\item{train_test}{Character to determine whether evaluation is run on training or test data
accepts "test", "train".}

\item{threshold}{vector of the same length as number of models chosen with custom
thresholds for model evaluation. for wsl.flex outputs the thresholds have to be labelled
with the same names provided to models}
}
\value{
an obejct of class 'wsl.evaluation'
}
\description{
Assess several model skill metrics for all models in a wsl.fit object. Currently
AUC, RMSE, TSS, PPV, Accuracy, and Cohen's Kappa are evaluated. Furthermore, the
threshold applied is returned.
}
\author{
Philipp Brun (main) with edits by Antoine Adde and Ben Black
}
