% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.finalisemodelspecifications.r
\name{lulcc.finalisemodelspecifications}
\alias{lulcc.finalisemodelspecifications}
\title{lulcc.finalisemodelspecifications}
\usage{
lulcc.finalisemodelspecifications(Model_specs_path, Param_grid_path)
}
\arguments{
\item{Model_specs_path}{Character, file path for table of model specifications}

\item{Param_grid_path}{Character, file path for grid of hyperparameters}
}
\description{
Enter optimal model specification and parameters for prediction
}
\author{
Ben Black
}
