% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.fitrf.r
\name{lulcc.fitrf}
\alias{lulcc.fitrf}
\title{define wsl.glm function}
\usage{
lulcc.fitrf(
  trans_result = numeric(),
  cov_data = data.frame(),
  replicatetype = character(),
  reps,
  strata = NA,
  save = FALSE,
  project = NA,
  path = NA,
  mod_args = list(),
  Downsampling = TRUE
)
}
\arguments{
\item{trans_result}{vector with presence/absence values}

\item{cov_data}{data.frame with environmental predictors}

\item{replicatetype}{(how) should replicates be generated? may be 'none', 'splitsample',
'cv' 'block-cv'}

\item{reps}{number of replicates}

\item{strata}{a numeric vector of the same length as observations with integers separating
cross validation replicates (used when replicatetype='block-cv')}

\item{save}{should the model be saved in a structured way? (not implemented yet)}

\item{project}{character indicating the name of the project within which the models are run
(later used to define saving directories)}

\item{path}{where to save? (not implemented yet)}

\item{mod_args}{list with elements of class 'multi.input' which specify models to be fitted
in lulcc.fitrf}

\item{step}{(for glms and gams only) should the models be updated with the step function?}

\item{mod_tag}{(not in wsl.flex) label for the current model}
}
\value{
Object of class wsl.fit including slots for meta info, testing data for
evaluation, and model objects
}
\description{
Re-sample data and fit RF models
}
\author{
Philipp Brun (main) with edits by Antoine Adde, adapted by Ben Black
}
