% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.listbybioregion.r
\name{lulcc.listbybioregion}
\alias{lulcc.listbybioregion}
\title{lulcc.listbybioregion}
\usage{
lulcc.listbybioregion(All_trans_list)
}
\description{
Seperate list of dataframes into a nested list according to Bioregion
}
\author{
Ben Black
}
