% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.modelstatscomparison.r
\name{lulcc.modelstatscomparison}
\alias{lulcc.modelstatscomparison}
\title{lulcc.modelstatscomparison}
\usage{
lulcc.modelstatscomparison(model_eval_summary, eval_metrics, grouping_var)
}
\arguments{
\item{model_eval_summary}{Dataframe summary of model performance}

\item{eval_metrics}{Vector, names of evaluation metrics to summarize by
can include: AUC, AUC.S, RMSE, Boyce, Score, threshold, Sensitivity,
Specificity, Accuracy, PPV, NPV, Jaccard, TSS Kappa, SEDI.}

\item{grouping_var}{Character, name of column to group results by}
}
\value{
List of model comparison results objects: stats test results, plots
}
\description{
Check assumptions and perform hypothesis testingfor differences in model eval metrics
between different model specification
}
\author{
Ben Black
}
