% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.preps.r
\name{lulcc.preps}
\alias{lulcc.preps}
\title{preparation function}
\usage{
lulcc.preps(env = parent.frame(), call)
}
\description{
Check input data, collect meta information, take care of data subsetting. Called by
model fitting functions.
}
\details{
Not to be called directly by the user
}
\author{
Philipp Brun adapted by Ben Black
}
