% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.spatprobmanipulation.r
\name{lulcc.spatprobmanipulation}
\alias{lulcc.spatprobmanipulation}
\title{lulcc.spatprobmanipulation}
\usage{
lulcc.spatprobmanipulation(
  Interventions,
  Scenario_ID,
  Raster_prob_values,
  Simulation_time_step
)
}
\arguments{
\item{Scenario_ID}{Chr, abbreviation of current scenario}

\item{Raster_prob_values}{Dataframe, predicted cellular transitions
probabilities with x,y and cell ID cols}

\item{Simulation_time_step}{Chr, current simultion time step}
}
\description{
function to manipulate spatial transition probabilities to represent
interventions/policies/management activities
Date: 17-03-2023
}
\author{
Ben Black
}
