% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.splitforcovselection.r
\name{lulcc.splitforcovselection}
\alias{lulcc.splitforcovselection}
\title{lulcc.splitforcovselection:}
\usage{
lulcc.splitforcovselection(trans_dataset, predictor_table)
}
\description{
TODO text
}
\author{
Ben Black
}
