% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.ceval.r
\name{pipe.ceval}
\alias{pipe.ceval}
\title{evaluate core function (peval)}
\usage{
pipe.ceval(f, pa, tesdat, crit, tre = numeric())
}
\arguments{
\item{f}{Vector of model predicted probabilities}

\item{pa}{Vector of response variable values}

\item{tesdat}{Dataframe of test data}

\item{crit}{which threshold criterion should be considered? Currently 'pp=op'
(predicted prevalence = observed prevalence), 'maxTSS' (threshold yielding maximum TSS),
and 'external' (thresholds manually supplied) are possible}

\item{tre}{Optional numeric threshold for seperating predictions}
}
\value{
Matrix of model eval metric names and results
}
\description{
Do the actual model evaluations; Not to be called directly by the user
}
\author{
Philipp Brun (main) with edits by Antoine Adde and Ben Black
}
