% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.df_or_rast.r
\name{pipe.df_or_rast}
\alias{pipe.df_or_rast}
\title{predict df or raster}
\usage{
pipe.df_or_rast(mod, nwdat, ...)
}
\description{
Predict to data.frame or raster; Not to be called directly by the user
}
\author{
Philipp Brun (main) with edits by Antoine Adde
}
