% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.prd.r
\name{pipe.prd}
\alias{pipe.prd}
\title{prediction-evaluation prediction function}
\usage{
pipe.prd(mod, tst)
}
\description{
Correctly feed the predict functions depending on model type (glm, gbm, maxent...)
}
\details{
Not to be called directly by the user
}
\author{
Philipp Brun (main) with edits by Antoine Adde
}
