% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preva_meta.r
\name{preva.meta}
\alias{preva.meta}
\title{prediction-evaluation meta.info function}
\usage{
preva.meta(env = parent.frame(), type = character())
}
\description{
Generate meta information for prediction and evaluation
}
\details{
Not to be called directly by the user
}
\author{
Philipp Brun
}
