% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_render_save_upload.r
\name{Quarto_render_save_upload}
\alias{Quarto_render_save_upload}
\title{Quarto_render_save_upload}
\usage{
Quarto_render_save_upload(Quarto_path, File_ext, Trackdown_auth_path)
}
\arguments{
\item{Quarto_path}{chr file path of Quarto document (.qmd) that you wish to save}

\item{File_ext}{chr either "html" or "pdf"}

\item{Trackdown_auth_path}{chr path to .json file for authentication for trackdown to
upload to Google drive}
}
\description{
Render a Quarto document locally, save it and then upload to Google drive using the
trackdown package. Prior to using you must follow the instructions on the trackdown
vignette for \href{https://claudiozandonella.github.io/trackdown/articles/oauth-client-configuration.html}{OAuth Client App Configuration}
}
\author{
Ben Black
}
