% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.neighbourhoodlayercreation.r
\name{random_pythagorean_matrix}
\alias{random_pythagorean_matrix}
\title{lulcc.neighbourhoodlayercreation}
\usage{
random_pythagorean_matrix(n, x, interpolation = "smooth", search = "random")
}
\description{
TODO text
}
\author{
Ben Black (lulcc function), Majid Shadman Roodposhti (matrix functions)
}
