% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.neighbourhoodlayercreation.r
\name{randomPythagorianMatrix}
\alias{randomPythagorianMatrix}
\title{lulcc.neighbourhoodlayercreation}
\usage{
randomPythagorianMatrix(n, x, interpolation = "smooth", search = "random")
}
\description{
TODO text
}
\author{
Ben Black (lulcc function), Majid Shadman Roodposhti (matrix functions)
}
