% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_prep.r
\name{region_prep}
\alias{region_prep}
\title{region_prep}
\usage{
region_prep()
}
\description{
Preparing Rasters of Bioregions in Switzerland to sub-divide data for regional
modelling This script is adapted from the process of Gerecke et al. 2019
}
\author{
Ben Black
}
