% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{setup}
\alias{setup}
\title{LULCC_CH setup: control table, config}
\usage{
setup(
  scenario_names = c("BAU", "EI-NAT", "EI-CUL", "EI-SOC", "GR-EX"),
  scenario_start = 2020L,
  scenario_end = 2060L,
  step_length = 5L,
  reps = 1L,
  simctrl_tbl_path = "tools/simulation_control.csv"
)
}
\description{
Creates / reads control table and config files
}
