% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcontrolprep.R
\name{simcontrolprep}
\alias{simcontrolprep}
\title{Create a simulation control table}
\usage{
simcontrolprep(
  Scenario_names = c("BAU", "EI-NAT", "EI-CUL", "EI-SOC", "GR-EX"),
  Scenario_start = 2020L,
  Scenario_end = 2060L,
  Step_length = 5L,
  reps = 1L
)
}
\arguments{
\item{Scenario_names}{char}

\item{Scenario_start}{1985 or later}

\item{Scenario_end}{2060 per default}

\item{Step_length}{in years}

\item{reps}{Number of runs for each simulation}
}
\description{
From master lulcc: how to prepare a sim control table
User enter start and end dates for the scenarios either enter a single number
value or a vector of values the same length as the number of scenarios earliest
possible model start time is 1985 and end time is 2060, simulations begin from
2020 and we have initially agreed to use 5 year time steps
}
