% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulcc.fitmodel.r
\name{stripGlm}
\alias{stripGlm}
\title{define wsl.glm function}
\usage{
stripGlm(cm)
}
\arguments{
\item{trans_result}{vector with presence/absence values}

\item{cov_data}{data.frame with environmental predictors}

\item{replicatetype}{(how) should replicates be generated? may be 'none', 'splitsample',
'cv' 'block-cv'}

\item{reps}{number of replicates}

\item{strata}{a numeric vector of the same length as observations with integers separating
cross validation replicates (used when replicatetype='block-cv')}

\item{project}{character indicating the name of the project within which the models are run
(later used to define saving directories)}

\item{path}{where to save? (not implemented yet)}

\item{step}{(for glms and gams only) should the models be updated with the step function?}

\item{mod_tag}{(not in wsl.flex) label for the current model}

\item{mod_args}{list with elements of class 'multi.input' which specify models to be fitted
in lulcc.fitrf}

\item{Downsampling}{Logical, should downsampling be implemented in RF? (TRUE/FALSE)}
}
\value{
Object of class wsl.fit including slots for meta info, testing data for
evaluation, and model objects
}
\description{
Re-sample data and fit RF models
}
\author{
Philipp Brun (main) with edits by Antoine Adde and Ben Black
}
