% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_fit_class.r
\docType{class}
\name{wsl.fit-class}
\alias{wsl.fit-class}
\alias{wsl.fit}
\title{Set wsl.fit Class}
\description{
An S4 class to store fitted objects
}
\section{Slots}{

\describe{
\item{\code{meta}}{A list with meta information}

\item{\code{tesdat}}{A list with held out data to be used for testing}

\item{\code{performance}}{fitted model objects}

\item{\code{call}}{the function call}
}}

