% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_zenodo_predictors.r
\name{fetch_zenodo_predictors}
\alias{fetch_zenodo_predictors}
\title{Get predictor data from Zenodo}
\usage{
fetch_zenodo_predictors(doi = "10.5281/zenodo.8263509")
}
\arguments{
\item{DOI}{which identifier to use}
}
\description{
This fetches predictor data from a very specifically structured Zenodo repo, namely
like \url{https://doi.org/10.5281/zenodo.8263509}
}
